added more detail upgrade of module
we specify that the module version should be upgraded. Customers don't know at which level the upgrade of the module needs to be done i.e. just at the cluster level or at the DB level as well. So just added that details and also a link on how to upgrade a module.