Changed the title of the IOPS panel
IOPS is confusing. This panel is only applicable for RedisOnFlash and relates to IOs. Customers who don't see any data and who expect the metrics to show instantaneous operations per second are wondering why the graph as no data. Changing the title to indicate that 1) it is for RoF, and 2) it is disk I/Os seems like a good change.