update replication.md
Changed the term 'node' to 'shard' to clarify that a) this is the data path that we're talking about and b) that this really is shards which are masters and slaves! Although a cluster does indeed have a 'master' node this node is in the control path, not the data path.