Update cluster-recovery.md
I think that the known limitation should be added to this page and moved out of the 5.4.10 release notes. This is not an issue that pertains to 5.4.10 and customers who are going to do a recovery are more likely to check the recovery documentation than the release notes.
I copied the text straight from the 5.4.10 release notes (and I left it there but I don't like duplicate information since it's hard to maintain. We should remove the next from the release notes IMO).