Add a link to the configuration page
It's not immediately obvious that you can customize redisInsight to change things like the default port. If I were a customer, I would look under 'troubleshooting' for how to do this. Therefore, I think a link should be added to the 'configuring RedisInsight' page. I've now had several customers ask if this was possible so I think this would be a great addition!