Update importing-data.md
In the import from a Redis server, fixed a typo indicating that a user should put RDB details to source DB details [see edited line].

Another point is that the section mentions "import a dataset from any publicly available Redis server" but this is not accurate.
As far as I know, import from a Redis Labs DB is not allowed. I'm not sure if there are other restrictions.
Need to check with PM to clarify that point.
I tried importing from both RS and Essentials (both publicly available) and I couldn't.