Update view-edit-cloud-account.md
Added a terraform module which automates all of this work, leveraging the newly released RedisCloud Terraform provider.

This work is incomplete, in the sense that the module needs to be moved from my personal Github account to a RedisLabs account, but is otherwise functional and tested. Who needs to move this module across and exactly where they must move it to is something we need to figure out. 

Furthermore, the module includes the two policy files - it would be good to automate the updating of this module when those two files are themselves updated.