Need to add a very important note about images version tag
I am not sure where to add this or how to make it as a "warning note", but it is very important that customers will use the exact same version tags as the ones we have on dockerhub. This might apply only to the main RS image, and not the others, but better make this a good practice.
We have has multiple customer issues because they used version tags that were custom and that the operator could not parse.