Update create-aa-database.md
About:  "If the API call fails, create a DNS alias that resolves your API hostname (`<api-hostname>`) to the IP address for the ingress controller's LoadBalancer."
I am not sure we can say this because the call could fail for many reasons (as simple as having used the wrong username and/or password. Instead of "create a DNS alias", maybe we could say something like "make sure that the API hostname" can be resolved properly and that you are using the right credentials".