Update monitor-performance.md
Screenshots in this page need to be updated to reflect the new UI.
The only change I made to the page was to add a line at the bottom '<<<NEED TO UPDATE SCREENSHOTS TO NEW UI>>>' just to indicate a change to the page. The line can be removed once the screenshots are updated.