Add flag that advises using the internal endpoint (#1737)
* Add flag that advises using the internal endpoint

While this documentation gives all of the relevant details on setting up VPC peering, it should also include a note (possibly highlighted or in a section that draws the eyes of the user) that the application connection string needs to be altered to point to the internal endpoint now. While this might seem obvious, there have been a couple of cases where customers enabled VPC peering, but didn't switchover the endpoints. This has led to these customers incurring large network costs due to the egress traffic traversing public internet. Again, this is in rare cases where there is high egress traffic, but it's easy enough to add the note and will likely prevent these future issues.

* Update content/rc/security/vpc-peering.md

Updating to reflect style and editorial guidelines.

Co-authored-by: Lance Leonard <lanceleonard@users.noreply.github.com>